refactoring eliza’s files to fit into a convention, moving abook and UrbanAirship test classes to a test directory
