+++ kifi-backend/modules/common/test/com/keepit/abook/ABookServiceClientTestModule.scala
@@ -1,0 +1,20 @@
+package com.keepit.abook
+
+import com.google.inject.{Provides, Singleton}
+import com.keepit.common.healthcheck.AirbrakeNotifier
+import com.keepit.common.net.HttpClient
+import com.keepit.common.zookeeper.ServiceDiscovery
+import com.keepit.common.service.ServiceType
+import play.api.Play._
+import net.codingwell.scalaguice.ScalaModule
+
+case class TestABookServiceClientModule() extends ABookServiceClientModule {
+  def configure() {}
+
+  @Singleton
+  @Provides
+  def ABookServiceClient(airbrakeNotifier: AirbrakeNotifier): ABookServiceClient = {
+    new FakeABookServiceClientImpl(airbrakeNotifier)
+  }
+
+}
