--- kifi-backend/modules/eliza/app/com/keepit/eliza/SomeElizaClass.scala
@@ -1,4 +1,0 @@
-
-class SomeElizaClass {
-    // code!
-}
