+++ kifi-backend/modules/eliza/app/com/keepit/eliza/controllers/mobile/MobileMessagingController.scala
@@ -1,6 +1,6 @@
 package com.keepit.eliza.controllers.mobile
 
-import com.keepit.eliza.controllers.internal.MessagingController
+import com.keepit.eliza.commanders.MessagingCommander
 import com.keepit.common.controller.{ElizaServiceController, MobileController, ActionAuthenticator}
 
 import play.api.libs.json.Json
@@ -9,12 +9,14 @@ import play.api.libs.concurrent.Execution.Implicits.defaultContext
 import com.google.inject.Inject
 
 
-class MobileMessagingController @Inject() (messagingController: MessagingController, actionAuthenticator: ActionAuthenticator)
-  extends MobileController(actionAuthenticator) with ElizaServiceController {
+class MobileMessagingController @Inject() (
+  messagingCommander: MessagingCommander,
+  actionAuthenticator: ActionAuthenticator
+  ) extends MobileController(actionAuthenticator) with ElizaServiceController {
 
   def getNotifications(howMany: Int) = AuthenticatedJsonAction { request =>
-    Async(messagingController.getLatestSendableNotifications(request.userId, howMany.toInt).map{ notices =>
-      val numUnreadUnmuted = messagingController.getUnreadUnmutedThreadCount(request.userId)
+    Async(messagingCommander.getLatestSendableNotifications(request.userId, howMany.toInt).map{ notices =>
+      val numUnreadUnmuted = messagingCommander.getUnreadUnmutedThreadCount(request.userId)
       Ok(Json.arr("notifications", notices, numUnreadUnmuted))
     })
   }
